+++ spec/shoulda/matchers/active_model/validate_numericality_of_matcher_spec.rb
@@ -236,6 +236,28 @@ describe Shoulda::Matchers::ActiveModel::ValidateNumericalityOfMatcher do
     end
   end
 
+  context 'with large numbers' do
+    it do
+      expect(validating_numericality(greater_than: 100_000))
+        .to matcher.is_greater_than(100_000)
+    end
+
+    it do
+      expect(validating_numericality(less_than: 100_000))
+        .to matcher.is_less_than(100_000)
+    end
+
+    it do
+      expect(validating_numericality(greater_than_or_equal_to: 100_000))
+        .to matcher.is_greater_than_or_equal_to(100_000)
+    end
+
+    it do
+      expect(validating_numericality(less_than_or_equal_to: 100_000))
+        .to matcher.is_less_than_or_equal_to(100_000)
+    end
+  end
+
   context 'with a custom validation message' do
     it 'accepts when the messages match' do
       expect(validating_numericality(message: 'custom')).